fix: consolidate legacy duplicate backend code trees
Backend had two parallel route/service trees: backend/routes + backend/
services (legacy) alongside backend/src/routes + backend/src/services
(canonical, wired up via src/index.ts and the dev/build scripts). Chose
backend/src/ as the single canonical tree and migrated every file the
legacy tree still actually used into it:

- services/{gmail-service,outlook-service,email-scanner,email-parser,
  service-categories,invoice.service,paystack}.ts and
  services/{subscription-classifier,subscription-creation}.js ->
  src/services/
- routes/integrations/{gmail,outlook,slack,classification-routes} ->
  src/routes/integrations/
- routes/csp-violations.js -> src/routes/csp-violations.ts, converted
  from CommonJS require/module.exports to the same ES import/export
  style as its new sibling route files, and fixed two pre-existing bugs
  found while moving it: it referenced a non-existent `authenticateToken`
  export (the real export is `authenticate`), and it was never mounted
  in src/index.ts even though client/app/api/csp-report/route.ts has
  always called POST /api/csp-violations expecting it to exist (now
  mounted)

Updated every import path this touched: src/index.ts, src/services/
email-rescan-service.ts, src/services/subscription-service.ts, three
backend test files, the root app/api/invoice-url/route.ts (Next.js
route consuming backend code via the @/ alias), and backend/tsconfig.json
(dropped the now-redundant explicit include entry for the file that's
naturally under src/** now). Verified every relative import this PR
touches resolves to a real file on disk via a throwaway Node script
(no TypeScript install needed/available in this environment).

backend/routes/ and backend/services/ no longer exist -- one canonical
tree remains. Updated backend/README.md's directory structure section
to match.

Note: backend/utils/ (a separate, smaller pre-src legacy utils tree,
still used by several of the moved files via relative imports) is left
untouched -- it wasn't named in this issue's evidence list and
consolidating it is a separate, similarly-scoped follow-up.
